requirements.txt
absl-py==0.7.1
aniso8601==7.0.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.8.0
blis==0.2.4
Bottleneck==1.2.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cvxpy==1.0.24
cycler==0.10.0
cymem==2.0.2
dill==0.3.0
ecos==2.0.7.post1
fancyimpute==0.5.3
fastai==1.0.57
fastprogress==0.1.21
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
knnimpute==0.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
multiprocess==0.70.8
murmurhash==1.0.2
numexpr==2.7.0
numpy==1.17.0
nvidia-ml-py3==7.352.0
osqp==0.5.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
pipenv==2018.11.26
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
pynvx==1.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scikit-surprise==1.0.6
scipy==1.3.1
scs==2.1.1.post2
seaborn==0.9.0
six==1.12.0
soupsieve==1.9.3
spacy==2.1.8
srsly==0.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc==7.0.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
wasabi==0.2.2
Werkzeug==0.15.4
wrapt==1.11.1
